+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
+### Added
+- BioNano Access server API - check projects, samples and fetch FSHD reports
 ### Changed
 - Updated igv.js to 2.15.6
 ### Fixed